+++ .gitignore
@@ -8,3 +8,4 @@ coverage
 coverage.json
 lcov.info
 dist
+lib
