+++ lerna.json
@@ -1,0 +1,4 @@
+{
+  "packages": ["packages/*"],
+  "version": "0.1.17"
+}
