+++ packages/ethereum-generator/src/@types/solc/index.d.ts
