+++ packages/ethereum-generator/src/template/AtomicPredicate.ts
