+++ packages/ethereum-generator/src/template/CompiledPredicate.ts
