+++ packages/ethereum-generator/src/template/DataTypes.ts
