+++ packages/ethereum-generator/src/template/UniversalAdjudicationContract.ts
