+++ packages/ethereum-generator/src/template/Utils.ts
