+++ packages/ethereum-generator/src/template/index.ts
