+++ packages/generator/src/CodeGenerator.ts
@@ -1,4 +1,4 @@
-import { CompiledPredicate } from '../transpiler'
+import { CompiledPredicate } from '@cryptoeconomicslab/ovm-transpiler'
 
 export interface CodeGenerator {
   generate(claimDefs: CompiledPredicate[]): Promise<string>
