+++ packages/parser/src/PropertyDef.ts
