+++ packages/solidity-generator/package.json
@@ -1,0 +1,55 @@
+{
+  "name": "@cryptoeconomicslab/ovm-solidity-generator",
+  "version": "0.0.1",
+  "description": "OVM transpiler",
+  "author": {
+    "name": "Cryptoeconomics Lab",
+    "url": "https://www.cryptoeconomicslab.com/"
+  },
+  "engines": {
+    "node": ">=10"
+  },
+  "license": "Apache-2.0",
+  "main": "lib/index.js",
+  "directories": {
+    "lib": "lib",
+    "test": "__tests__"
+  },
+  "files": [
+    "lib"
+  ],
+  "publishConfig": {
+    "access": "public"
+  },
+  "scripts": {
+    "build": "node preprocessor.js && tsc --build",
+    "build:check": "tsc --noEmit",
+    "test": "jest --passWithNoTests",
+    "lint": "eslint --ext ts,tsx src",
+    "doc": "typedoc --out docs src",
+    "clean": "rm -rf lib docs"
+  },
+  "jest": {
+    "transform": {
+      "^.+\\.tsx?$": "ts-jest"
+    },
+    "testMatch": [
+      "**/__tests__/**/?(*.)+(spec|test).(ts|js)",
+      "**/?(*.)+(spec|test).ts"
+    ],
+    "moduleFileExtensions": [
+      "js",
+      "ts"
+    ],
+    "globals": {
+      "ts-jest": {
+        "tsConfig": "tsconfig.json"
+      }
+    }
+  },
+  "dependencies": {
+    "ejs": "^3.0.1",
+    "@cryptoeconomicslab/ovm-transpiler": "^0.0.1",
+    "@cryptoeconomicslab/ovm-generator": "^0.0.1"
+  }
+}
