+++ packages/solidity-generator/preprocessor.js
@@ -2,14 +2,13 @@ const fs = require('fs')
 const path = require('path')
 
 const list = [
-  './src/parser/chamber.peg',
-  './src/generator/solidity/sol.ejs',
-  './src/generator/solidity/decide.ejs',
-  './src/generator/solidity/getChild.ejs',
-  './src/generator/solidity/constructProperty.ejs',
-  './src/generator/solidity/constructInputs.ejs',
-  './src/generator/solidity/constructInput.ejs',
-  './src/generator/solidity/decideProperty.ejs'
+  './src/sol.ejs',
+  './src/decide.ejs',
+  './src/getChild.ejs',
+  './src/constructProperty.ejs',
+  './src/constructInputs.ejs',
+  './src/constructInput.ejs',
+  './src/decideProperty.ejs'
 ]
 
 list.forEach(file => {
