+++ packages/solidity-generator/src/.gitignore
