+++ packages/solidity-generator/src/constructInput.ejs
