+++ packages/solidity-generator/src/constructInputs.ejs
