+++ packages/solidity-generator/src/constructProperty.ejs
