+++ packages/solidity-generator/src/decide.ejs
