+++ packages/solidity-generator/src/decideProperty.ejs
