+++ packages/solidity-generator/src/getChild.ejs
