+++ packages/solidity-generator/src/sol.ejs
