+++ packages/transpiler/src/QuantifierTranslater.test.ts
@@ -1,5 +1,5 @@
 import { translateQuantifier } from './QuantifierTranslater'
-import { PropertyDef } from '../parser/PropertyDef'
+import { PropertyDef } from '@cryptoeconomicslab/ovm-parser/lib/PropertyDef'
 
 describe('QuantifierTranslater', () => {
   beforeEach(async () => {})
