--- src/generator/index.ts
@@ -1,3 +1,0 @@
-export * from './CodeGenerator'
-export * from './ethereum'
-export * from './solidity'
